requirements.txt
# DAOS-RFF: Domain-Adaptive Open-Set RF Fingerprinting
# Requirements for Lightning AI L4 GPU

# Core ML Framework
torch>=2.0.0
pytorch-lightning>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# Data Processing
numpy>=1.24.0
scipy>=1.10.0
h5py>=3.8.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization & Logging
matplotlib>=3.7.0
tensorboard>=2.12.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.2.0
requests>=2.28.0

# Signal Processing (for STFT spectrograms)
librosa>=0.10.0
